moved statement factories to their own folder
