+++ src/game/gdl/statement/factory/GeneratorFactory.groovy
@@ -1,6 +1,11 @@
-package game.gdl.statement
+package game.gdl.statement.factory
 
 import game.GameContextInfo
+import game.gdl.statement.GDLStatement
+import game.gdl.statement.GeneratorStatement
+import game.gdl.statement.SimpleStatement
+import game.gdl.statement.TokenUser
+import game.gdl.statement.Tokens
 
 
 /**
