+++ src/game/gdl/statement/factory/StatementFactory.groovy
@@ -1,6 +1,9 @@
-package game.gdl.statement
+package game.gdl.statement.factory
 
 import game.GameContextInfo
+import game.gdl.statement.GDLStatement
+import game.gdl.statement.GeneratorStatement
+import game.gdl.statement.factory.GeneratorFactory
 
 /**
  * @author Lawrence Thatcher
